templates: Add an option for full description on builtin_log_compact
For a new user, it is not clear how to view the full commit
message/description of a change with `jj log`. There is a
`builtin_log_detailed` template, which is closer to the default log
format that git users are probably used to. However, this has two
(minor) issues:

1. **Discoverability**. Speaking from my own subjective experience, this
   is not easy to find for a new user, for what is likely a common
   operation. Reading `jj log -h` or `jj help log` does not offer any
   clues that `builtin_log_detailed` exists (if one is lucky, one might
   run `jj log -T` with no argument and see `builtin_log_detailed` in
   the list). In my case, I did not discover this until I cloned the
   `jj` repo and found this listed in the `templates.toml` file.

2. **Information density**. `builtin_log_detailed` lives up to its name
   and is quite detailed. Ideally, there should be a way to view the
   entire commit message/description in the log, but with the "normal"
   log output format. The `builtin_log_detailed` template shows the full
   commit/change ID (without highlighting the "shortest prefix") and
   puts each piece of information on a separate line.

This fix this, add an option `full_description` to display the full
description on `builtin_log_compact`. To not break compatibility, this
is set to `false` by default.

The user can set it to true on the config like this:
```
templates.log = builtin_log_compact(true)
```

Fixes: #3688